refactor types directory
