+++ src/types/index.ts
@@ -1,21 +1,2 @@
-import type { User, Session, Profile } from './supabase';
-import type {
-    MovieResults,
-    MovieDetailsData,
-    TvResults,
-    TvDetailsData,
-    ShowProviders,
-    ShowData,
-} from './tmdb';
-
-export type {
-    User,
-    Session,
-    Profile,
-    MovieResults,
-    MovieDetailsData,
-    TvResults,
-    TvDetailsData,
-    ShowProviders,
-    ShowData,
-};
+export type * from './supabase';
+export type * from './tmdb';
