+++ src/types/tmdb/index.ts
@@ -1,0 +1,3 @@
+export type * from './movie';
+export type * from './tv';
+export type * from './show';
